+++ appinfo/routes.php
@@ -17,6 +17,8 @@ return [
 
 		['name' => 'openAiAPI#getModels', 'url' => '/models', 'verb' => 'GET'],
 		['name' => 'openAiAPI#getPromptHistory', 'url' => '/prompts', 'verb' => 'GET'],
+		['name' => 'openAiAPI#getUserQuotaInfo', 'url' => '/quota-info', 'verb' => 'GET'],
+		['name' => 'openAiAPI#getAdminQuotaInfo', 'url' => '/admin-quota-info', 'verb' => 'GET'],
 		['name' => 'openAiAPI#clearPromptHistory', 'url' => '/clear-prompt-history', 'verb' => 'POST'],
 		['name' => 'openAiAPI#createCompletion', 'url' => '/completions', 'verb' => 'POST'],
 		['name' => 'openAiAPI#createImage', 'url' => '/images/generations', 'verb' => 'POST'],